created WrapGPUDST.cpp with conditional includes for FFTW, CUDA and HIP
